Move sudoku_cell into a separate file. Fmt and lint
